+++ src/main/java/backend/wal/wal/todaywal/application/service/RegisterReservationTodayWalService.java
@@ -1,13 +1,13 @@
-package backend.wal.wal.todaywal.domain.service;
+package backend.wal.wal.todaywal.application.service;
 
 import backend.wal.wal.todaywal.application.port.RegisterReservationTodayWalUseCase;
-import backend.wal.wal.common.domain.aggregate.WalCategoryType;
-import backend.wal.wal.common.domain.aggregate.WalTimeType;
-import backend.wal.wal.todaywal.domain.aggregate.TodayWal;
 import backend.wal.wal.todaywal.domain.repository.TodayWalRepository;
-import backend.wal.support.annotation.DomainService;
+import backend.wal.wal.todaywal.domain.aggregate.TodayWal;
+import backend.wal.wal.common.domain.WalCategoryType;
+import backend.wal.wal.common.domain.WalTimeType;
+import backend.wal.support.annotation.AppService;
 
-@DomainService
+@AppService
 public class RegisterReservationTodayWalService implements RegisterReservationTodayWalUseCase {
 
     private final TodayWalRepository todayWalRepository;
