+++ src/main/java/backend/wal/wal/todaywal/application/service/UpdateTodayWalShowStatusService.java
@@ -1,13 +1,13 @@
-package backend.wal.wal.todaywal.domain.service;
+package backend.wal.wal.todaywal.application.service;
 
 import backend.wal.wal.todaywal.application.port.UpdateTodayWalShowStatusUseCase;
-import backend.wal.wal.todaywal.domain.aggregate.TodayWal;
 import backend.wal.wal.todaywal.domain.repository.TodayWalRepository;
+import backend.wal.wal.todaywal.domain.aggregate.TodayWal;
 import backend.wal.wal.todaywal.exception.NotFoundTodayWalException;
-import backend.wal.support.annotation.DomainService;
+import backend.wal.support.annotation.AppService;
 import org.springframework.transaction.annotation.Transactional;
 
-@DomainService
+@AppService
 public class UpdateTodayWalShowStatusService implements UpdateTodayWalShowStatusUseCase {
 
     private final TodayWalRepository todayWalRepository;
