+++ src/main/java/backend/wal/wal/todaywal/domain/aggregate/ShowStatus.java
@@ -1,4 +1,4 @@
-package backend.wal.onboard.todaywal.domain.aggregate;
+package backend.wal.wal.todaywal.domain.aggregate;
 
 public enum ShowStatus {
     OPEN,
